requirements.txt
awsebcli
boto
boto3
celery
coverage
emoji
flask
flask_restful
gunicorn
nose
pillow
requests==2.6.1 # Needed for awsebcli.
stripe
tabulate
twilio
uwsgi
